+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.597] - 2021-10-21
 ### Add
-- Add Input Number
+- Add Progress Circular
 
 ## [0.1.596] - 2021-10-15
 ### Add